Rename wp.js to wpapi.js
