+++ package.json
@@ -6,7 +6,7 @@
   "name": "wpapi",
   "version": "0.9.3",
   "description": "A client for interacting with the WordPress REST API",
-  "main": "wp.js",
+  "main": "wpapi.js",
   "repository": {
     "type": "git",
     "url": "https://github.com/wp-api/node-wpapi.git"
@@ -29,8 +29,8 @@
     "docs": "grunt docs",
     "jekyll": "node bin/jekyll",
     "release-docs": "node build/scripts/release-docs",
-    "jshint": "jshint --reporter=node_modules/jshint-stylish/index.js Gruntfile.js wp.js bin build lib tests",
-    "jscs": "jscs Gruntfile.js wp.js bin build lib tests --reporter node_modules/jscs-stylish/jscs-stylish.js",
+    "jshint": "jshint --reporter=node_modules/jshint-stylish/index.js Gruntfile.js wpapi.js bin build lib tests",
+    "jscs": "jscs Gruntfile.js wpapi.js bin build lib tests --reporter node_modules/jscs-stylish/jscs-stylish.js",
     "lint": "npm run jshint && npm run jscs || true",
     "mocha": "_mocha tests --recursive --reporter=nyan",
     "watch": "grunt watch",
